requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
pandas
pypdf
faiss-cpu